client: support associated account, sender create, (#6)
and create attribution.

Add support for using the associated account on creation,
attribution for account creation (i.e. a memo with the
configured app index in create transactions), and sender
create semantics for SendPayment().

Note: there is no support for sender create in SubmitEarnBatch,
as the transaction payload size limits us to a batch size of 2
in the worst case. This is not worth the complexity, and so
it is recommended that users use SendPayment with type Earn
instead.
GitOrigin-RevId: 91298f85ea863a3691fe5f059086cf3c578b38f0